feature/organize-files - Merge ticket and ticketrouter for better reading. Merge user and userrouter for better reading
