+++ internal/api/ticket.go
@@ -6,10 +6,13 @@ import (
 	"time"
 
 	"github.com/gin-gonic/gin"
+	"github.com/go-playground/validator/v10"
+	"github/moura95/meli-api/internal/service"
 	"github/moura95/meli-api/internal/util"
 	"github/moura95/meli-api/pkg/errors"
 	"github/moura95/meli-api/pkg/ginx"
 	"github/moura95/meli-api/pkg/jsonplaceholder"
+	"go.uber.org/zap"
 )
 
 type ticketResponse struct {
@@ -301,4 +304,30 @@ func (t *TicketRouter) hardDelete(ctx *gin.Context) {
 	}
 
 	ctx.JSON(http.StatusOK, ginx.SuccessResponse("Ok"))
+}
+
+type ITicket interface {
+	SetupTicketRoute(routers *gin.RouterGroup)
+}
+
+type TicketRouter struct {
+	service  service.TicketService
+	logger   *zap.SugaredLogger
+	validate validator.Validate
+}
+
+func NewTicketRouter(s service.TicketService, log *zap.SugaredLogger) *TicketRouter {
+	return &TicketRouter{
+		service:  s,
+		logger:   log,
+		validate: *validator.New(),
+	}
+}
+
+func (t *TicketRouter) SetupTicketRoute(routers *gin.RouterGroup) {
+	routers.GET("/tickets", t.list)
+	routers.GET("/tickets/:id", t.get)
+	routers.DELETE("/tickets/:id", t.hardDelete)
+	routers.POST("/tickets", t.create)
+	routers.PATCH("/tickets/:id", t.update)
 }
