+++ internal/api/user.go
@@ -7,6 +7,7 @@ import (
 	"github.com/gin-gonic/gin"
 	"github/moura95/meli-api/pkg/ginx"
 	"github/moura95/meli-api/pkg/jsonplaceholder"
+	"go.uber.org/zap"
 )
 
 func (u *UserRouter) list(ctx *gin.Context) {
@@ -45,4 +46,23 @@ func (u *UserRouter) get(ctx *gin.Context) {
 	}
 
 	ctx.JSON(http.StatusOK, ginx.SuccessResponse(response))
+}
+
+type IUser interface {
+	SetupUserRoute(routers *gin.RouterGroup)
+}
+
+type UserRouter struct {
+	logger *zap.SugaredLogger
+}
+
+func NewUserRouter(log *zap.SugaredLogger) *UserRouter {
+	return &UserRouter{
+		logger: log,
+	}
+}
+
+func (t *UserRouter) SetupUserRoute(routers *gin.RouterGroup) {
+	routers.GET("/users", t.list)
+	routers.GET("/users/:id", t.get)
 }
